requirements.txt
Django>=5.2.1
djangorestframework>=3.14.0
celery>=5.2.7
redis>=4.5.1
graphene-django>=3.2.0
djangorestframework-simplejwt>=5.2.2
python-dotenv>=1.0.0